Support memory bank buffer allocation without cl_ext pointer.
This enhances current support without requiring single master
connection for each CU.  A CU can now be connected to multiple memory
banks, device buffer allocation happens as part of clSetKernelArg,
where enough information is available to compute which bank to use.

No changes required for exisitng code.

To use the enhancemet, the host code should call clSetKernelArg
immediately after clCreateBuffer and before clEnqueueMigrateMemObjects.
This implies that a pipelined or ping-pong application should allocate
a working set of buffers and kernel objects upfront and reuse these
objects in the loop body withput calling clCreateKernel and
clSetKernelArg in the body of the loop.