Fix ubuntu packaging when libraries are linked against version libraries
The packager for DEBIAN does a dependency check on binaries.  If any
dependent library cannot be found in library search path, then the
dpkg-shlibdeps errors out.  This can be fixed in a couple of ways

- Link libraries with RPATH set
- Set LD_LIBRARY_PATH prior to invoking the dpkg-build
- Change access mode on .so to executable
- Disable SHLIB dependency check

First is probably the better solution, but need more investigation.
Second is not feasible since it is cmake that invokes the packager and
we don't really know location it puts the libraries.  Thirds is just
not good because .so should not be execuable.  Fourth is simplest and
what was done in this commit.