modified xbb device IDs (#174)
Hi Lizhi, Hem,

We made some slight updates to device IDs for the U* Boards. Please see the updated table here:

http://confluence.xilinx.com/display/XSW/DSA+Tracker

Note we’re reserved 4 PFs per board although today our DSAs only use 2. MGMT PF will be PF0 and User PF will be PF1 on these boards. Also for the MFG Device ID the driver should not pair when it sees this device ID because this is not a DSA and just the MFG Image which is used to recover the flash. Xbflash will have these special device IDs in its code so it recognizes the device IDs.