Do not throw bad_alloc if kernel argument bank check fails
Kernel argument check is called during set_argument, but may not be
able to do allocation check if the device context has multiple
devices.  At the time the kernel argument is set the targeted device
is unknown.  In other words, the argument cannot be allocated before
an enqueue operation where the targeted device is known.