Multiprocess work around for xclbin locking and icap reset of kds
Fixes problem related to xclbin locking and icap reset of kds as part of AXI reset.  The AXI reset is performed when a process tries to download an xclbin (icap_download_bitstream_axlf) *and* that xclbin is unused (first download or already downloaded but unlocked).

- P1 downloads xclbin calling xocl_read_axlf_helper
-- not locked so resets AXI => resets KDS
-- aquires xclbin lock (ref=1)
-- releases xclbin lock (ref=0)
- P1 configures scheduler
-- This does not mess with xclbin locking, which remains ref=0
-- Configure command enters KDS command queue
-- P1 waits in execWait for configure command to complete
- P2 downloads xclbin
-- not locked so resets AXI => resets KDS
-- The pending P1 configure command is ‘stale’ from KDSs perspective, this is the point of reset, so the command is removed
-- acquires xclbin lock (ref=1)
-- releases xclbin lock (ref=0)

If P1 is slow getting serviced by KDS, P1 will never complete, it hangs in xclExecWait for configure command.

The work-around is to not release lock after xclbin download, meaning that all processes must agree to same xclbin.