Lock xclbin prior to using KDS
Fixes problem related to xclbin locking and icap reset of kds as part of AXI reset.  The AXI reset is performed when a process tries to download an xclbin (icap_download_bitstream_axlf) *and* that xclbin is unused (first download or already downloaded but unlocked).

- P1 downloads xclbin calling xocl_read_axlf_helper
-- not locked so resets AXI => resets KDS
-- aquires xclbin lock (ref=1)
-- releases xclbin lock (ref=0)
- P1 configures scheduler
-- This does not mess with xclbin locking, which remains ref=0
-- Configure command enters KDS command queue
-- P1 waits in execWait for configure command to complete
- P2 downloads xclbin
-- not locked so resets AXI => resets KDS
-- The pending P1 configure command is ‘stale’ from KDSs perspective, this is the point of reset, so the command is removed
-- acquires xclbin lock (ref=1)
-- releases xclbin lock (ref=0)

If P1 is slow getting serviced by KDS, P1 will never complete, it hangs in xclExecWait for configure command.

The fix in this PR is to lock the xclbin for a client context prior to submitting a command to KDS.  The xclbin is locked only if necessary, no repeated attempts to lock an already locked xclbin.

Error check that context locks the expected xclbin, not some xclbin swapped in by another process.