kernel arg connectivity deduction
Multiple changes to support clSetKernelArg deducing device buffer
memory connectitivity.

Kernel is constructed with all device CUs based on loaded program.
The kernel CUs are dynamically filtered based on connectivity
requirements of kernel arguments.

The kernel CUs that remain after filtering are sticky in the sense
that when reusing a kernel object, only the prior filtered resulting
CUs are avaiable, if other CUs are required, a new kernel object should
be constructed.

This commit also does away with the memory extention where
{kernel,argidx} is specified, Though still supported, clCreateBuffer
simply calls clSetKernelArg under the hood when the the buffer carries
the now obsolete extension.

The deduced memory connectivity of a cl_mem buffer is stored as a data
member on the memory object.  This is used for the cases where the
targeted device is unknown until first clEnqueue operation involving
the buffer.