Close possible race condition in init of shm db and mutex
It is possible that a 2nd process mapping in the shared
memory may do so before the mutex is initialized by the
originating process.  This fix eliminates that possible
error condition and introduces an up-to 500ms loop condition
in the event that process 2 is launched before process 1
fully initializes the shared memory region.

Signed-off-by: Jeffrey Mouroux <jmouroux@xilinx.com>