Use ERT for CU polling in dataflow mode (#1079)
In dataflow mode KDS configures, starts, and checks CU status
directly.

If KDS is configured with ert=false, then CUs are polled for
completion in a busy loop.

If KDS is configured with ert=true (and KDS is in dataflow mode), then
ERT is used to poll CUs for completion and notify KDS when one
completes.  If KDS is not busy attempting to start commands, then it
will be waiting for ERT notification.  The ERT command queue is
divided into CU slots, where each slot corresponds to a CU.  KDS
requests ERT polling of a CU by writing AP_START or AP_CONTINUE into
the CQ slot corresponding to the CU.

Until xclbin meta data advertises dataflow kernels, sdaccel.ini can be
used to trigger dataflow mode by setting Runtime.dataflow=true.  Note
that together with ert=true (the default) the requires updated ERT
firmware part of the PR that supports polling of CUS.