Support CU selection with clCreateKernel (#1082)
* Support CU selection with clCreateKernel

cl_kernel clCreateKernel(program, kernel_name, err) now supports an
extention to the kernel_name such that in can include the compute unit
that should be used.

The format of the kernel_name argument is "kernel:{cu1,cu2,...}" where
"kernel" is the name of the kernel as normal and "cu1" through "cu_n"
are the names of the compute units to use.  It is strictly enforced
that the CU list is enclosed in a comma separated list without any
spaces between the CU names, commas, and curly braces.  It is also
strictly enforced that the list must follow a ":" separating a legal
kernel name.  Invalid CU names are currently ignored without warning,
but error is thrown if no CUs matched.

* Fix compile error on 18.04