In dataflow mode up the limit of cmds sent to scheduler
For NDRange kernels (non-datflow) with multiple workgroups, we limit
the number of workgroups (cmds) sent to scheduler in order to prevent
filling CQ with cmds all waiting for same CUs.

For dataflow kernels the limit is increased since each CU now
pipelines cmds up to the pipeline depth.  This depth is unknown so the
fix in far from ideal but good for 2019.1 where we are not limited by
a HW CQ in dataflow mode.