Add back disabled kernels to compute units in userspace
Userspace needs to contruct kernel objects in order to create the
stream objects, the kernel objects must contain the streaming CUs
otherwise connectivity (route and flow ids) of the stream can't be
determined.

How the disabled kernels are handled in userspace and kernel space is
not very clean.  I think this can be simplified potentially if the
stream userspace code is rewritten.

User space creates compute units for streaming kernels, but these
compute units cannot be used for scheduling.  User space ensures that
these compute units are assigned indidces after regular AXI-lite CUs.
In kernel space all CUs are passed to scheduler in configure command
but the streaming CUs will not be used for scheduling commands.
Context creation where CUs are reserved will error out if user space
attempts to reserve a streaming CU and command validation will fail if
a command targets a streaming CU.

Disallowing clients from reserving a streaming CUs is not correct and
must be fixed in the future when we decide the mechanism for reserving
a streaming CU.

CR:1028256,1028200