xma: xmahw_hal: Change ddr_banks default array initializer to zero
In some platforms like the MPSoC platforms, the xclbin may not use
real fabric/pl based kernels, but instead may use the ARM cores (PS)
via soft-kernels. Which means every xma plugin may or may not have
a real kernel that it uses. Which means there may not be a real
DDR bank that XMA should try to parse from the xclbin. Since MPSoC based
platforms typically have only one DDR bank, change the default array
initializer to be 0 instead of -1.

A ddr bank initializer of -1 bears the consequence of setting P2P bits to true
when allocating BOs which is not supported on the MPSoC platform.

Signed-off-by: Rohit Athavale <rohit.athavale@xilinx.com>