2019.1: skd: Change process name to match soft kernel name
Commit 9270ab0 was merged into XRT/master by PR-1512. This commit applies it to XRT/2019.1
https://github.com/Xilinx/XRT/pull/1512

The commit tries to change the process name to match the soft kernel name.
This proves to be useful when trying to debug the performance of specific
soft kernels, unique names aids in tracking in these details.

Note: This still only impacts the name seen in top, ps still remains the same process name.

Signed-off-by: Rohit Athavale <rathaval@xilinx.com>
Reviewed-by: Larry Liu <yliu@xilinx.com>