skd: Change process name to match soft kernel name
The commit tries to change the process name to match the soft kernel name.
This proves to be useful when trying to debug the performance of specific
soft kernels, unique names aids in tracking in these details.

Moved to the child process execution path immediately after the fork() as
suggested by Larry.

Note: This still only impacts the name seen in top, ps still remains the same

Signed-off-by: Rohit Athavale <rathaval@xilinx.com>