Small refactor and fix for clEnqueueReadBufferRect
There was no setIfZero causing this function to work incorrectly in 
certain circumstances, it would copy to the first section of the host 
pointer repeatedly. This results in invalid outputs where the final 
section of the source buffer would be copied to the first section of the 
source buffer and the rest of the values would be unmodified.

Signed-off-by: Andrew Gozillon <andrew.gozillon@yahoo.com>