Refactor and Fix for clEnqueueCopyBufferRect/clEnqueueReadBufferRect (#1541)
* Incorrect extent in bytes calculation for clEnqueueCopyBufferRect

This calculation seems to be a little off, it currently prohibits 
copying 2D regions from buffers that are the same size as the 2D buffer. 
The value it computes is always larger than the buffer.

* Small refactor of clEnqueueCopyBufferRect

* Small refactor and fix for clEnqueueReadBufferRect

There was no setIfZero causing this function to work incorrectly in 
certain circumstances, it would copy to the first section of the host 
pointer repeatedly. This results in invalid outputs where the final 
section of the source buffer would be copied to the first section of the 
source buffer and the rest of the values would be unmodified.

Signed-off-by: Andrew Gozillon <andrew.gozillon@yahoo.com>