Removed the use of __FUNCTION__ (#2336)
The XclBinExceptionUtil exception class was using the __FUNCTION__ macro to set the default value for _file in the constructure.  Unfortunately doing so caused the Windows compiler to fail.  The code was refactored to no longer depend on this macro.
Note
----
If the user wishes to have the function name to be part of the thrown exception, they would need to add this value during the exception class creation.