xrt-next.h: Use structs for profile related members & routines
Several routines in this header refer to ProfileResults, however it was declared as
a simple struct. Without the typedef specifier it will not be resolved when using
an application that eventually needs compile with xclhal2.h

v1 : typedef usage for each data member
v2 : pure struct based declarations as suggested by Soren.

This commit changes the profile related function definitions to use struct * instead
of typedefs and allows applications that use those headers to compile.

Signed-off-by: Rohit Athavale <rohit.athavale@xilinx.com>