Enhance clGetDeviceInfo to return BDF for a device
This is applicable only for HW flow, emulation defaults to unknown.

Current implementation is not clean, it uses a sysfs path to parse for
BDF.  In future this should go through device classes used by xbutil
and xbmgmt, but more restructuring is needed in order to do this.

Prepare device_linux for accessing more than sysfs data in query
function.