Clear cache execbufs when device is close
This was yet another regression from #3053, where devices are closed
by OCL when no longer in use.

For example if a cl_context is released, then the devices used by the
context are closed (if noone else have them opened).  However the
xrt::device object remains alive.

If host application constructs a new context from same devices, then
they are opened again, this time with a different xclDeviceHandle.

The execbuffer cache used by OCL caches execbuf for xrt::device
objects, so the previous cached execbufs must be purged when/if the
underlying device is closed so that new execbufs can be constructed
for proper xclDeviceHandle.