CR-1062504 multiple thread hang waiting for commands to finish (#3236)
This is specific to AP_CTRL_CHAIN and ERT assisted polling for CU
completion.  KDS can be ahead of ERT in that it doesn't have to wait
for an interrupt form ERT before determining that a CU has completed.
When KDS gets ahead, there could be a pending interrupt for one of CUs
that KDS itself determined as completed.  Without acknowledging the
interrupt, ERT cannot send future interrupts and KDS can in certain
cases go into a waiting state when no further commands arrive and none
appear to have completed.

The fix is to force one extra iteration of the scheduler after
transitioning new commands from submitted to running.  This way the
interrupt register gets cleared by the code that checks if running
commands are done.