CR-1067817 add new condition in clock scaling feature
The new bit 30 (in 0xB20 reg) set to '1' indicates that the currently
loaded SC firmware does not support clock throttling.
XRT can set bit 29 (Feature present) if the ULP metadata indicates that
hardware support for throttling is loaded, but should not set bit 28
(Feature enabled) if the new bit 30 is set to '1'.
A dmesg report should be generated to indicate that runtime clock
throttling is not supported if the user tries to enable clock throttling
when bit 30 is set to '1'.

Signed-off-by: Rajkumar Rampelli <rajkumar@xilinx.com>