XRT-604 xmc: skip loading xmc firmware during xrt installation
New proposed shell(NSPR-7555) will embeds CMC.elf in the bitfile and
it will separates out the ERT & CMC reset, so XRT can avoid
resetting the CMC but continue to load/reset ERT as it is doing today.
Added required changes in XMC driver to support this change.

This change is backward compatible since the register field which
used here to determine if xmc is in bitfile will have 0 value in old shells.
And, the new shell which embeds cmc.elf in bitfile will have 1 value.

Signed-off-by: Rajkumar Rampelli <rajkumar@xilinx.com>