C++ API rst documentation
This PR adds support for C++ API documentation.  This is done through
doxygen and a sphinx plugin 'breathe'.  The documetation step in XRT,
runs doxygen on C++ headers to genetated XML data consumed and
formatted by sphinx with the breathe plugin.

This PR also separates documentation generation from building.  By
default the build.sh script no longer generates documentation.  This
is done because the necessary version of sphinx and plugins are not
available on all platforms.  In order to generate documentation, the
build.sh script must be invoked with the '-docs' option, e.g.

% build.sh -docs

Documentation generation will work if and only if sphinx-build and
breathe are installed along with kernel-doc.