Support find_package on windows (#4508)
This pull request modifies #4505 by adding back the EXPORT option for
the component install target.  This is necessary for windows where the
link library is released by the development component install target.

The new EXPORT is called xrt-target-dev, it is released by
findpackage.cmake and is magically included by xrt-target.cmake,
because xrt-target-dev matches the GLOB pattern used.  Presumably
xrt-target-dev.cmake could be explicitly included by the config file
(xrt.fp.in), but currently this is not necessary. xrt-targets-dev is
conditionally installed for win32 as it empty on Linux which fails
with some version of CMake.

Use find_package to located XRT libraries in low level XRT tests.