CR-1087412 xrt API has lower IOPS result vs. xcl API
Revamp how kernel commands are submitted for execution and checked for
completion.

Command completion now supports both poll and push mode with poll
corresponding to calling xrt::run::wait and push correspondng to the
original command monitor thread.

Turns out that push is only required if a command completion callback
is installed, which primarily is true only in enqueued operations such
as in OpenCL.

When no callbacks are installed, commands are submitted for execution
bypassing the execution monitor.  These commands are marked complete
only when they are explicitly queried by calling the command wait
function.  This mode of execution has almost no overhead compared to
the xcl shim APIs. The only overhead is to ensure thread safety such
that multiple threads, indirectly through the APIs, can call exec wait
without the risk of loosing command completions.  This in itself has
been a long outstanding issue with xclExecWait.

This PR is a breaking change.  It is no longer possible to install
command callbacks after a command has been submitted for execution.
An exception is thrown if this is attempted.  The work-around is to
create the xrt::run object, install the callbacks, and then start the
execution.  This ensures that push mode is used when the command is
submitted for execution.

Enhance noop emulation to insert completion delay