Rename export target for development content
Rename xrt-targets-dev to xrt-dev-targets to avoid naming conflict
with exported release target xrt-targets. Ideally no release targets
should be exported but until XRT splits into a release and development
package the release content appears to must be exported in order for
find_package to work on Linux.

This is work-in-progress.