CR-1091258 U200 Crash: return code None
Hang in execwait similar to what was fixed in #4878.  On this branch
the fix is less intrusive and just removes the exec_wait call for
scheduled KDMA execbuf.

Eliminating the exec_wait implies busy wait for KDMA completion, which
is not a problem given that the old execution monitor is used only for
KDMA right now.  In 2021.1 the old execution monitor can be completely
eliminated.