Update static boost build for later versions of boost
% boost.sh -prefix <dir>/boost-1.75.0 boost-1.75.0
% env XRT_BOOST_INSTALL=<dir>/boost-1.75.0/xrt build/build.sh -opt

Above ensures that XRT shared libaries are linked with static
boost libraries and all boost symbols are resolved.

This PR fixes a few build errors when using latest version of boost.
It also adds support for static linking of boost on aarch64 (arm64).
Finally, the PR ensures that local version of boost can be used when
static tool build of XRT is enabled.