Bug Fix: xbmgmt examine command didn't check for existance of directory (#5128)
The search algorithm for platforms examines a collection of root platform directions.  Depending on the compiler, there is a chance where checking of the directory existance is being performed prior to determine if the path was a directory or file.  When this occurs an exception is thrown.

This fix assures that doesn't happen.

Co-authored-by: Stephen Rozum <stephenr@xilinx.com>