CR-1100135 Error calling err = q.enqueueTask(kernel2, NULL, &event1)
Fix memory bank grouping such that banks with same base address and size
share the same encoded index. This is a regression per
changes Jan 4 that switched OpenCL to use native APIs.  In particular
commit e9160b11 neglected to merge identical banks.

This would be much simpler if the xclbin itself had merged identical
banks into groups as it does for explicit grouping of for example
HBMs.