Fix regex matching of kernel name (#5489)
* Fix regex matching of kernel name

If clCreateKernel or xrt::kernel objects are contructed with a kernel
name that is a substring of other kernel names, then matching fails
because of missing ":" in dynamically created regex.

Add missing ":" seperator in regex "(kernelname):(.*)".

* Update python bindings for xrt::xclbin::kernel access

And fix 22_verify.py to use kernel name rather than ip name.  The
ipname is treated as a kernel name and no longer matches because ":"
is inserted after then specified name prior to matching against
IP_LAYOUT kernel instances.