Addressed Windows Debug build compiler warnings (#5716)
The Windows Debug Build of XBMain.c produced a C4702 compiler warning
that indicated code was unreachable.  Further examination of the code
suggests that that issue is associated with the MS compiler and not the
actual code. Never the less, the code was refactored to remove the
"break" statement (the code that was being complained about) from the
for loop.

Co-authored-by: Stephen Rozum <stephenr@xilinx.com>