Ensure schema compatibile xrt::device::get_info return values
In order to ensure that json formatted return values of
xrt::device::get_info conforms to the same schema in future
versions of XRT, the call is capturing the version of XRT used
when the host application is built.

The compiled version is passed to implementation of get_info() which
can then ensure schema compatibility as necessary.

* No user visible API change, just call templated public get_info(param)
  as normal.  This templated function in turn calls new overloaded
  private get_info().

* Add overloaded private get_info(param, abi) where the abi argument
  is really the version of XRT used when the host application was
  built.

* Pass the ABI (the application version of XRT) to json_str(…) and use
  in the future when / if new schema.

* Leave existing private get_info(param) untouched so that 2021.1 host
  application that reference this symbol can continue to find it in
  new versions of XRT.  Remove after a few years.