azure: parse xclbin and check if it is valid to send
Parse xclbin in the azure mpd plugin before sending it to mgmt and check if it is valid
xclbin to send.
If the xclbin has BITSTREAM section then it is either DCP xclbin or bit.xclbin (full bitstream), both
these xclbins are not valid to program.

So, adding a check in azure mpd plugin to verify the same.

https://dev.azure.com/AzureXilinx/XRT%20Collaboration/_workitems/edit/84/

Signed-off-by: Rajkumar Rampelli <rajkumar@xilinx.com>