CR-1114287 XRT] ERROR: No such xclbin handle: Invalid argument
Handles created by C-APIs are mapped to a managed shared
implementation object that is reference counted.  When handle is
closed, the implementation can still be alive.

This PR solves a race condition when inserting and accessing the map.

A handle map was created to ensure exclusive access.

The bug has been present since inception, e.g., the C-APIs are not
exception safe without using a mutex protected handle map.