Add domain to kernel cu index (#5994)
In preparation for xrt::kernel support for pskernel, the compute unit
index has to differentiate between pl and ps kernels.  Rather than
adhoc add 128 (maxcu) to offset the ps kernel index, this PR
introduces the concept of a CU domain.

The CU domain is owned by xocl kernel module.  The domain is different
for a PL kernel CU (cu) and a PS kernel CU (scu), and it is encoded in
the index returned by the driver through sysfs.

In userspace, the encoded index is used when opening a context (shared
or exclusive) on the CU.  The index within a domain (the domain index)
is used in the cumask of the command packet sent to KDS.  The domain
index is obtained by stripping off the domain from the encoded index.

With this change in place, xrt::kernel will be remain more or less
unchanged in the code path that processes and opens compute units
associated with the kernel.

While transitioning to using encoded cu indices, the core device class
adds a new data member which is a mapping from CU name to encoded
index. Following PRs will make use of this mapping when constructing
xrt::kernel objects and opening contexts, but in this PR the mapping
is unused.