Update xrt::kernel implementation to use domain/index for compute unit
This PR builds on #5994 to leverage the CU domain.

Context management is per domain, so the grunt of the work in this PR
is adding an extra level of indexing to structures that capture and
manage context opened and closed using xclOpenContext and
xclCloseContext.  The extra level of indexing is basically a mapping
from domain to contexts acquired and released within that domain.

The CU domain index remains consecutive and in the range from 0 to
max_cus (128).  Kernel execution uses the domain index for the
command cumask.

The risk in this PR is that native xrt::kernel APIs have switched to
use the mapping from CU name to encoded CU index introduced in

The standard OpenCL HW tests have been validated against the changes
in this PR.