Fix PS kernel instance name to match up with IP_LAYOUT
Modify instance name to match up with xclbin convention where
instance names are numbered scu_0..n as in kernel_name:scu_[0..n]
for kernel_name. It is important that the instance name matches up with the
expectations of core XRT, which is based off of the instance names in
the IP_LAYOUT.

The instance names are created in user space XRT from the data
obtained from kernel driver, which doesn't use the IP_LAYOUT section
for PS kernels and in fact is not concerned with instance name beyond
the kernel name itself.

This PR uses a using a knm -> idx map, which allows the driver to
assign internal indeces in any order it likes and sysfs node does not
have to list all kernel instances in order of kernel names.

There is no real risk with this PR as only new #6010 needs the
instance names matching with IP_LAYOUT, this PR just enables support
for multiple different PS kernels each with any number of instances.