Vitis-3888 Automatic buffer migration using M2M (#6083)
Support automatic local buffer creation to allow read access from
buffers allocated in memory bank that is incompatible with kernel
connectivity.

When global buffer is used as argument to kernel, a connectivity check
ensures that the global buffer is compatible with the kernel.  With
this PR an incompatible global buffer is cloned with same properties
as original but allocated in the bank that is compatible with the
kernel.

The local buffer is associated with the original global buffer from a
lifetime point of view, but there is no data integrity management
between the clone and its source, meaning that once cloned any changes
to the source buffer are not reflected in clones if any.

Most pratical use of this new cloning feature is when using outputs of
one kernel as input to some other kernel.  The cloning (if necessary)
is expensive as it involves DMA of all source data to the clone.

A simple illustrative example exercising the feature has been added to
the XRT native API tests.