CR-1116820 Buffer export handle is not closed
When exporting a buffer object, the export handle (file descriptor
linux) is not closed automatically so the underlying drm buffer is
not freed even when the exporting buffer object is deleted.

In this PR, the export handle is managed by the exporting buffer
object implmentation and lifetime is tied to that of the buffer
object.  When the buffer object is destructed, the export handle (if
any) is closed.  The changes in this PR apply to both native XRT APIs
and OpenCL.

There are two semantic changes in this PR that should be noted:

First, the export handle of a buffer object is now stored with the
exporting buffer object.  If a buffer object is exported multiple
times, the same export handle (fd) is returned.

Second, the export handle lifetime is tied to the exporting buffer
which cannot be closed until after the export handle is no longer
needed.

The changes were tested per the associated CR.

Documentation probably should reflect the lifetime requirement
mentioned above.