Issue #6027 Update Vitis default paths in scripts
Issues #6027 raises concerns about magically defaulting Vitis paths in
build.sh.

Defaulting the paths will not change, but they are updated to more
recent versions and warning is issued if path cannot be found.
Removing the paths and relying on internal users managing the setting
of XILINX_VITIS is not feasible.  The scripts are primarily for
convienience of internal users.

Risk is minimal resulting from path update in build.sh and
preservation of XILINX_XRT in run.sh.

This PR closes #6027.