Remove boost::any dependency from xrt_device.h
Default building of XT to use std=c++17 and use std::any in place of
boost::any in the xrt::device exposed get_info() API.  This removes
the need to install boost-dev when using native XRT APIs.

Support backwards compatibility with host code compiled with earlier
versions of XRT.  If host application was compiled against earlier
version of XRT it references boost::any get_info symbols.  These
symbols are retained and old binaries will continue to work with new
XRT.  Also new binaries that are built with -std=c++14 will use
boost::any and require that boost-dev is installed.

The boost::any version of get_info() is hereby deprecated and will
be removed in the future in per deprecation policy.