VITIS-2660 Async programming support with Native C++ API
This PR implements generic C++ support for concurrent asynchronous
execution of commands enqueued into queues.

Within a single queue commands executes synchronously, but between
queues execution is asynchronous.

A single queue executes commands asynchronously with respect to the
enqueing thread.

When a command is enqueued the caller receives an event (future) which
becomes valid when the operations has completed.  The event can be
enqueued into multiple different queues in order to synchronize these
different queues with the completion of the corresponding command.

The event can also be used to retrieve the return value of a command
once it has completed execution.

The implementation has no dependencies on XRT itself.