WIP: Support multiple xclbins
This PR is first step in user space supporting loading of
multiple xclbins.

Adds the infrastructure for caching multiple xclbins in
xrt_core::device along with their slot id.  For flows that do not
support multiple xclbins, the infrastructure is still used, but
defaults to caching one xclbin only.

Re-factored when the cached xclbins are refreshed. The idea is that
refreshing can happen at any time.  First, regular host loaded xclbins
causes a refresh of the cache.  Second, during opening of a CU
context, it is possible the driver loads an existing xclbin into a new
slot, and if so, then cached slots must be refreshed.  The refactoring
decouples refreshing of the cache from the host loading of xclbin.

The changes have been tested through regular Alveo hardware regression
testing.