Support xrt::hw_context
New concept to represent hardware resources (columns, regions, PS
processes) used by an xclbin.  A hardware context is used by an
application to control where execution takes place.  In a multi-xclbin
scenario, the hardware context is specific to one xclbin even if that
xclbin is loaded multiple times.

The feature is optional for legacy applications / platforms.

Constructed from a device and an xclbin uuid that has been registered
with the device (driver)

A context is 1-1 with a slot index. A resource resolver assigns a
hardware context to hardware resources, maybe sharing, or fails.
On success, the driver assigns a unique slot index to the context.

A kernel object is created from a hardware context. CU context
creation no longer needs EAGAIN as opening of CU context always
succeeds if the hardware context was successfully created. The rather
complicated code for EAGAIN per #6421 has been removed in this PR.

A hardware context can be shared between threads. A process
can construct multiple hardware contexts.