CR-1130511 Eliminate padding at the end of mailbox messages
The drivers have been using sizeof(struct xcl_mailbox_req) to help
them size the buffers for mailbox messages, adding to this size the
length of any message content.  However, these calculations create
buffers that are larger than necessary, because the compiler pads
the structure for alignment purposes, and the message content is
not appended beyond that padding - it is instead inserted at the
data member of the structure.  This arrangement has the effect of
causing the buffers allocated to be larger than necessary, with
the padding shifted to the end of the buffer.  This patch fixes
the buffer sizing calculations to be aware of this behaviour and
eliminates the surplus storage.