Use hwctx when in IP context manager.
This is continuation of changes in #6727 and #6731.  Use of
xrt::hw_context will move all the way to xclOpenContextByName, but
changes are done in small incremental steps.

This PR changes the context manager to return the IP index upon
successfully opening the IP.  This simplifies how ip_context
gets the IP index.