Vitis-6101 Support read and write directly from/to device buffer
To avoid of host memcpy with the usage of xclSynBO(), Vitis AI runtime
(VART) currently leverages xclUnmgdPwrite() and xclUnmgdPread() to
perform PCIe DMA data transfer between host and device so as to
achieve better performance.

This PR extends the semantics of xclUnmgdPwrite/read into the native
xrt::bo APIs.  The xcl APIs are deprecated and should not be used.

Extend xrt::bo::read() and xrt::bo::write() to write directly to
device buffer if buffer object has not host backing storage.