Support command submission through hardware queues
Add internal hardware queue concept to make command submission
specific to a hardware context.  A hardware queue is an abstraction
that binds command submission to a specific hardware context.

The PR replaces high level use of the device specific command
submission with queue based command submission. The changes in this PR
are exclusively to XRT core functionality, at the lowest level, the
execbuf and execwait protocol remains unchanged and is still only
device specific.  Later PR will propagate the command queue to new
shim level APIs and also provide shim API for constructing the KMD
counterpart to the hardware queue if supported.

The hardware queue is not exposed to end-user applications and usage
is strictly limitted to xrt::kernel execution.  The xrt::kernel is
constructed from a hardware context and the hardware queue is created
internally and associated with the context.