Add experimental support for QoS specification in hwctx
Allow applications to create a hwctx specifying the necessary QoS.
This is *very* preliminary experimental support subject to change w/o
warning.

QoS for a hwctx is supposed to guide the resource solver to find
available HW resources sufficient to statisfy the specified QoS or
fail if no resources are available.  For platforms that support
multiple programmable regions (slots), the QoS can be used to
determine if an existing slot can be shared with another context or if
a new region (slot) needs to be programmed.

QoS is a set of metrics defined as a key-value pair represented as
std::map.  The accepted keys are communicated internally in this
version and their value type (represented as uint32_t) is also
internally documented.

The QoS is bound to evolve as more platform drivers support our
concept of xclbin sharing per QoS specifications, so view this PR as
work-in-progress trial feature until we determine exactly what metrics
are needed can be supported.

This PR changes the type of the QoS parameter and adds a second
parameter for legacy support of CU shared/exclusive access mode.  The
vast number of files changed is due to the extra access mode parameter
and the type changes of the QoS paramter. The QoS type is aliased in
one place so that future change to some more intelligent
representation should trigger ripple effects.

The changes has been tested with standard hw regression tests that
exercise legacy OpenCL test cases.