HW context specific group id for buffer allocation
Add API to get the memory group id associated with a hardware context.

For some platforms, buffer allocation can be specific to a hardware
context when the buffer is not specifically used as a kernel argument.

This PR add xrt::hw_context::get_memory_group_id() which returns an
encoded group id flag which can be used with xrt::bo constructors.