Fixed new-delete mismatches
Allocating an array type `T[]` using new, assigning to a non-array
pointer `*T` and then deleting (or freeing) the pointer could lead to
problems.  Instead of fixing the pointer type I used `make_unique()`[1]
which creates both the data and the pointer in one step. This avoids
duplicating the type both in the pointer and data.

`make_shared()` does not support array types in C++17[2], so I left
`unique_shared` and fixed the type.

[1]:
https://www.open-std.org/jtc1/sc22/wg21/docs/papers/2017/n4659.pdf#subsubsection.23.11.1.4
[2]:
https://www.open-std.org/jtc1/sc22/wg21/docs/papers/2017/n4659.pdf#paragraph.23.11.2.2.6