CR-1129189 Throw exception on abnormal command completion
New xrt::run::wait2() API was added to throw an exception when command
completes abnormally (state different from ERT_CMD_STATE_COMPLETED).

The existing run::wait() API cannot be changed without breaking ABI,
so new function had to be added.

The new wait2() API can be given an optional timeout value in which
case the API returns std::cv_status::timeout on the API timeout, or
std::cv_status::no_timeout on successful command completion.

An xrt::exception base class was added and a derived
xrt::run::command_error expception defined.  This was done so that the
abnormal command state can be captured if an exception is thrown by
the wait2() API.  The exception class is insulated, which may seem
like an overkill, but it to facilitate otherwise ABI breaking changes.